+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 ### Added
 - this changelog, markdown description content type tag for PyPI
 - use [Semantic Versioning](https://semver.org/spec/v2.0.0.html)
+- read back camera eye/target, light position, color and r/u/v
 
 ## 0.1.2.5 - 2019-04-16
 Two weeks and some steps from the initial release. Starting changelog w.r.t. this point.